+++ src/routes/(repl)/+layout.svelte
@@ -1,16 +1,13 @@
 <script lang="ts">
-	import { browser, dev } from '$app/environment';
 	import { afterNavigate, beforeNavigate } from '$app/navigation';
-	import { PUBLIC_SAVE_IN_LOCAL_STORAGE_NAME } from '$env/static/public';
 	import CommandRunner from '$lib/command_runner/CommandRunner.svelte';
 	import { commands } from '$lib/command_runner/commands';
 	import Credits from '$lib/components/Credits.svelte';
+	import Dialog from '$lib/components/Dialog.svelte';
 	import { is_repl_to_save, repl_id, repl_name } from '$lib/stores/repl_id_store';
 	import { webcontainer } from '$lib/webcontainer';
-	import { decompressFromEncodedURIComponent } from 'lz-string';
 	import type { LayoutData } from './$types';
-	import Dialog from '$lib/components/Dialog.svelte';
-	import { parse } from '$lib/components/parsers';
+	import { handle_load_files } from './handle_load_files';
 
 	export let data: LayoutData;
 
@@ -28,46 +25,9 @@
 	afterNavigate(async () => {
 		if (fix_for_double_after) return;
 		fix_for_double_after = true;
-		// if there's a data.promises.github_repo this means we
-		// are loading a github repo as project
-		if (data.promises?.github_repo) {
-			loading_github_repo = true;
-			webcontainer.set_file_system(
-				await data.promises.github_repo.then((github_content) => {
-					loading_github_repo = false;
-					return github_content;
-				})
-			);
-		} else {
-			// try to get the project from local storage and then delete it
-			const stored_project = window.localStorage.getItem(PUBLIC_SAVE_IN_LOCAL_STORAGE_NAME);
-			if (stored_project !== null) {
-				try {
-					const project = parse(stored_project);
-					await webcontainer.set_file_system(project);
-				} catch (e) {
-					/* empty */
-				}
-				window.localStorage.removeItem(PUBLIC_SAVE_IN_LOCAL_STORAGE_NAME);
-			} else {
-				// check if there's the hash
-				const hash = window.location.hash.substring(1);
-				if (hash) {
-					const url_search_params = new URLSearchParams(hash);
-					const code = url_search_params.get('code');
-					if (code) {
-						const project = decompressFromEncodedURIComponent(code);
-						try {
-							const to_mount = parse(project);
-							await webcontainer.set_file_system(to_mount);
-						} catch (e) {
-							/* empty */
-						}
-					}
-				}
-			}
-		}
-
+		await handle_load_files(data, (value) => {
+			loading_github_repo = value;
+		});
 		webcontainer.init().then(() => webcontainer.mount_files());
 		// for debugging
 		(window as any).wc = webcontainer;
