+++ src/routes/[github=github]/[...repo]/+server.ts
@@ -7,6 +7,7 @@ export const GET: RequestHandler = async ({ params }) => {
 		/(?<owner>[^/]+)\/(?<repo>[^/]+)\/?(?:tree\/(?<branch>[^/]+))?(?<path>\/.*)?/;
 	const info = params.repo.match(github_regex)?.groups;
 	const search_params = new URLSearchParams();
+	search_params.set('provider', 'github');
 	for (const group in info) {
 		if (info[group]) {
 			search_params.set(group, info[group]);
